requirements.txt
-i https://pypi.org/simple
apscheduler==3.6.3
cachetools==4.2.2 ; python_version ~= '3.5'
certifi==2022.9.24 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
idna==3.4 ; python_version >= '3.5'
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.28.1
setuptools==65.5.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.1 ; python_version >= '3.5'
tzdata==2022.6 ; python_version >= '3.6'
tzlocal==4.2 ; python_version >= '3.6'
urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
